PSY-951: split MarkdownContent + dynamic-import MarkdownEditor and dnd-kit to evict vendor libs from the shared chunk
