+++ frontend/features/collections/components/CollectionCard.tsx
@@ -22,7 +22,7 @@ import { Badge } from '@/components/ui/badge'
 import { UserAttribution } from '@/components/shared'
 import { formatRelativeTime } from '@/lib/formatRelativeTime'
 import { getEntityTypeLabel, type Collection } from '../types'
-import { MarkdownContent } from './MarkdownEditor'
+import { MarkdownContent } from './MarkdownContent'
 import { CollectionCoverImage } from './CollectionCoverImage'
 import { useLikeCollection, useUnlikeCollection } from '../hooks'
 import { useAuthContext } from '@/lib/context/AuthContext'
